# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Firma API
  description: API za upravljanje firmama i zaposlenima
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: Lokalni razvojni server

paths:
  /employees:
    get:
      summary: Vraća sve zaposlene
      operationId: getAllEmployees
      responses:
        '200':
          description: Lista svih zaposlenih
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
    post:
      summary: Kreira novog zaposlenog
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Zaposleni je uspešno kreiran

  /employees/{id}:
    get:
      summary: Vraća zaposlenog prema ID-u
      operationId: getEmployeeById
      parameters:
        - in: path
          name: id
          required: true
          description: ID zaposlenog
          schema:
            type: integer
      responses:
        '200':
          description: Podaci o zaposlenom
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Zaposleni nije pronađen
    put:
      summary: Ažurira podatke o zaposlenom
      operationId: updateEmployee
      parameters:
        - in: path
          name: id
          required: true
          description: ID zaposlenog
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Zaposleni je uspešno ažuriran
    delete:
      summary: Briše zaposlenog
      operationId: deleteEmployee
      parameters:
        - in: path
          name: id
          required: true
          description: ID zaposlenog
          schema:
            type: integer
      responses:
        '204':
          description: Zaposleni je uspešno obrisan
        '404':
          description: Zaposleni nije pronađen

  /employees/company/{companyId}:
    get:
      summary: Vraća sve zaposlene za određenu firmu
      operationId: getEmployeesByCompanyId
      parameters:
        - in: path
          name: companyId
          required: true
          description: ID firme
          schema:
            type: integer
      responses:
        '200':
          description: Lista zaposlenih u firmi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'

  /companies:
    get:
      summary: Vraća sve firme
      operationId: getAllCompanies
      responses:
        '200':
          description: Lista svih firmi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
    post:
      summary: Kreira novu firmu
      operationId: createCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        '201':
          description: Firma je uspešno kreirana

  /companies/{id}:
    get:
      summary: Vraća firmu prema ID-u
      operationId: getCompanyById
      parameters:
        - in: path
          name: id
          required: true
          description: ID firme
          schema:
            type: integer
      responses:
        '200':
          description: Podaci o firmi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '404':
          description: Firma nije pronađena
    put:
      summary: Ažurira podatke o firmi
      operationId: updateCompany
      parameters:
        - in: path
          name: id
          required: true
          description: ID firme
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        '200':
          description: Firma je uspešno ažurirana
    delete:
      summary: Briše firmu
      operationId: deleteCompany
      parameters:
        - in: path
          name: id
          required: true
          description: ID firme
          schema:
            type: integer
      responses:
        '204':
          description: Firma je uspešno obrisana
        '404':
          description: Firma nije pronađena

components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: integer
          description: ID zaposlenog
        name:
          type: string
          description: Ime zaposlenog
        position:
          type: string
          description: Pozicija zaposlenog
      required:
        - id
        - name
        - position

    Company:
      type: object
      properties:
        id:
          type: integer
          description: ID firme
        name:
          type: string
          description: Naziv firme
        address:
          type: string
          description: Adresa firme
      required:
        - id
        - name